Remove extensions file's suffix
