+++ lib/utils/extensions/theme_data.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+extension ThemeDataExtensions on ThemeData {
+  Color get onMain => switch (brightness) {
+        Brightness.light => colorScheme.onSurface,
+        Brightness.dark => colorScheme.onInverseSurface,
+      };
+}
