--- lib/utils/extensions/theme_data_extensions.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-extension ThemeDataExtensions on ThemeData {
-  Color get onMain => switch (brightness) {
-        Brightness.light => colorScheme.onSurface,
-        Brightness.dark => colorScheme.onInverseSurface,
-      };
-}
