--- lib/utils/extensions/tonal_palette_extensions.dart
@@ -1,6 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:material_color_utilities/material_color_utilities.dart';
-
-extension ToColor on TonalPalette {
-  Color getColor(int tone) => Color(get(tone));
-}
